feat: allow custom container component and more customization
By default, Autolink will continue to behave as before: output will be wrapped in a single Text node
and non-Autolink props will be passed through to that node. To allow for more customization, you can
now specify a custom component, for example View, to use as the container instead of Text using the
`component` prop. Any non-Autolink props will be passed through to this component instead, and all
text within is wrapped with Text components as required by RN. Additionally, the new `linkProps` and
`textProps` props allow you to pass any props to links or text components, respectively. And the new
`renderText` prop allows you to completely customize how text is wrapped in the output.

BREAKING CHANGE: Non-Autolink props are no longer passed to links. Only styles supplied to
`linkStyle` and props supplied to `linkProps` are used when rendering links. You are still free to
use `renderLink` to fully customize link rendering.

closes #48